Add re-usable base test class
